+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.1.1
 - 'movePCtoCursor' back in right-click menu.
+- "z88dkMapFile" parameter: better support for z88dk labels.
 
 # 1.1.0
 - zsim: